Moved user views into user module
