+++ src/modules/user/views/user_settings/components/DeleteAccount.vue
