+++ src/modules/user/views/user_settings/components/ProfileEditor.vue
