+++ src/modules/user/views/user_show/components/sidebar.vue
