+++ src/modules/user/views/user_show/components/tech.vue
